+++ src/main/java/org/individualproject/business/PaymentDetailsService.java
@@ -3,7 +3,11 @@ package org.individualproject.business;
 import org.individualproject.business.converter.PaymentDetailsConverter;
 import org.individualproject.business.converter.UserConverter;
 import org.individualproject.business.exception.InvalidExcursionDataException;
+import org.individualproject.business.exception.NotFoundException;
+import org.individualproject.business.exception.UnauthorizedDataAccessException;
+import org.individualproject.configuration.security.token.AccessToken;
 import org.individualproject.domain.*;
+import org.individualproject.domain.enums.UserRole;
 import org.individualproject.persistence.PaymentDetailsRepository;
 import org.individualproject.persistence.entity.PaymentDetailsEntity;
 import org.individualproject.persistence.entity.UserEntity;
@@ -17,6 +21,7 @@ import java.util.Optional;
 @Service
 public class PaymentDetailsService {
     private PaymentDetailsRepository paymentDetailsRepository;
+    private AccessToken requestAccessToken;
     @Autowired
     public PaymentDetailsService(PaymentDetailsRepository paymentDRepository){
         this.paymentDetailsRepository = paymentDRepository;
@@ -26,8 +31,22 @@ public class PaymentDetailsService {
         return PaymentDetailsConverter.mapToDomainList(paymentDetailsEntities);
     }
     public Optional<PaymentDetails> getPaymentDetails(Long id) {
-        Optional<PaymentDetailsEntity> paymentDetailsEntities = paymentDetailsRepository.findById(id);
-        return paymentDetailsEntities.map(PaymentDetailsConverter::mapToDomain);
+
+        Optional<PaymentDetailsEntity> paymentDetailsOptional = paymentDetailsRepository.findById(id);
+
+        if (paymentDetailsOptional.isPresent()) {
+            PaymentDetailsEntity paymentDetailsEntity = paymentDetailsOptional.get();
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name())) {
+                if (!requestAccessToken.getUserID().equals(paymentDetailsEntity.getUser().getId())) {
+                    throw new UnauthorizedDataAccessException("UNAUTHORIZED_ACCESS");
+                }
+            }
+
+            return paymentDetailsOptional.map(PaymentDetailsConverter::mapToDomain);
+        } else {
+            throw new NotFoundException("Payment details not found.");
+        }
     }
 
     public PaymentDetails createPaymentDetails(CreatePaymentDetailsRequest request){
@@ -49,10 +68,19 @@ public class PaymentDetailsService {
     }
 
     public boolean deletePaymentDetails(Long id) {
-        try {
+        Optional<PaymentDetailsEntity> paymentDetailsOptional = paymentDetailsRepository.findById(id);
+        if (paymentDetailsOptional.isPresent()) {
+            PaymentDetailsEntity paymentDetailsEntity = paymentDetailsOptional.get();
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name())) {
+                if (!requestAccessToken.getUserID().equals(paymentDetailsEntity.getUser().getId())) {
+                    throw new UnauthorizedDataAccessException("UNAUTHORIZED_ACCESS");
+                }
+            }
+
             paymentDetailsRepository.deleteById(id);
             return true;
-        } catch (EmptyResultDataAccessException e) {
+        } else {
             return false;
         }
     }
@@ -62,6 +90,12 @@ public class PaymentDetailsService {
         Optional<PaymentDetailsEntity> optionalPaymentDetails = paymentDetailsRepository.findById(request.getId());
         if (optionalPaymentDetails.isPresent()) {
             PaymentDetailsEntity existingPaymentDetails = optionalPaymentDetails.get();
+
+            if (!requestAccessToken.hasRole(UserRole.TRAVELAGENCY.name())) {
+                if (!requestAccessToken.getUserID().equals(existingPaymentDetails.getUser().getId())) {
+                    throw new UnauthorizedDataAccessException("USER_ID_NOT_FROM_LOGGED_IN_USER");
+                }
+            }
             existingPaymentDetails.setId(request.getId());
             existingPaymentDetails.setUser(userEntity);
             existingPaymentDetails.setCvv(request.getCvv());
